feat: Implement input validation in get_user_input function
This change introduces input validation within the get_user_input function
to ensure the Bunkrr Album URL is provided and in a valid format.

The validate_input decorator is added to perform input validation within the get_user_input function.
It includes checks to ensure the Bunkrr Album URL is not empty and
has a valid URL format using the validate_url function from the validators module.

If the URL input is empty or has an invalid format, a ValueError is raised to prompt correct input from the user.

Changes:
- Added validate_input decorator to enforce input validation in get_user_input function.
- Implemented checks to ensure the Bunkrr Album URL is not empty and has a valid URL format using validate_url function.
- Raised ValueError for invalid or empty URL input to request correct input from the user.

Signed-off-by: Ahmad Thoriq Najahi <najahi@zephyrus.id>